docs/search_engine_src/core/requirements_current.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.7.1
asttokens==2.4.1
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.40.26
botocore==1.40.26
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cobble==0.1.4
contourpy==1.3.2
couchbase==4.3.3
crypto==1.4.1
cryptography==43.0.1
cycler==0.12.1
decompyle3==3.9.2
decorator==5.1.1
distro==1.9.0
dnspython==2.8.0
dxf2svg==0.1.4
ecdsa==0.19.1
elastic-transport==8.15.1
elasticsearch==8.15.1
elasticsearch-dsl==8.15.1
email-validator==2.3.0
et-xmlfile==1.1.0
executing==2.1.0
ezdxf==0.8.8
fastapi==0.104.1
firebase-admin==6.5.0
Flask==2.3.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
fonttools==4.58.5
frozenlist==1.8.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.2.4
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
influxdb==5.3.2
iniconfig==2.0.0
ipython==8.28.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.0
mammoth==1.11.0
mariadb==1.1.11
markdown-it-py==4.0.0
MarkupSafe==3.0.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.7.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
Naked==0.1.32
-e git+https://github.com/wizice/neomoncli@0e6e475bf959681a1b8d5725e8ecc5139093ca9c#egg=neodevcli
-e git+https://github.com/wizice/neomoncli@0e6e475bf959681a1b8d5725e8ecc5139093ca9c#egg=neomoncli
-e git+https://github.com/wizice/neosis_www@f7b927c65fb09ae72d2995d122f184adf0f2b641#egg=neosiscli&subdirectory=cli
netifaces==0.11.0
networkx==2.6.3
numpy==2.1.2
openai==1.52.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
playwright==1.55.0
pluggy==1.5.0
polyline==2.0.3
prompt_toolkit==3.0.48
propcache==0.4.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycel==1.0b30
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.9.0
pymssql==2.3.5
PyMySQL==1.1.1
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
qrcode==8.2
redis==5.1.1
regex==2024.9.11
reportlab==4.4.5
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
scipy==1.14.1
seaborn==0.13.2
selenium==4.33.0
setuptools==75.2.0
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spark-parser==1.8.9
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
tabulate==0.9.0
telegram==0.0.1
tiktoken==0.8.0
tqdm==4.66.5
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.17.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uncompyle6==3.9.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
-e git+https://github.com/wizice/wizcli@e8424c1d6a45d55a40ee20d2c033cf33b65430f0#egg=wizcli3
wsproto==1.2.0
WTForms==3.1.2
xdis==6.1.2
xlsxwriter==3.2.9
xmltodict==0.14.2
yarl==1.22.0
zeep==4.3.1
